+++ anchor/programs/games/src/instructions/store/sale/client.rs
@@ -3,14 +3,16 @@ use crate::constants::{COLLECTOR_SEED, STORE_SEED};
 use crate::instructions::sale::utils::*;
 use anchor_lang::prelude::*;
 use anchor_spl::token::{Mint, Token, TokenAccount};
+use crate::instructions::StorePackage;
 
-pub fn trader(ctx: &Context<StoreSale>, amount: u64) -> Result<()> {
-    let sale = Sale {
+pub fn trader(ctx: Context<StoreSale>) -> Result<()> {
+    let mut sale = Sale {
         tollkeeper: ctx.accounts.tollkeeper.clone(),
         bump: ctx.bumps.tollkeeper,
         collector: ctx.accounts.collector.clone(),
         trader: ctx.accounts.trader.clone(),
         store: ctx.accounts.store.clone(),
+        package: ctx.accounts.package.clone(),
         payer: ctx.accounts.payer.clone(),
         receiver: ctx.accounts.receiver.clone(),
         feed: ctx.accounts.feed.clone(),
@@ -19,13 +21,15 @@ pub fn trader(ctx: &Context<StoreSale>, amount: u64) -> Result<()> {
         token_program: ctx.accounts.token_program.clone(),
     };
 
-    sale.charge(amount)?;
-    sale.transfer(amount)?;
+    sale.charge()?;
+    sale.transfer()?;
+    ctx.accounts.package.increment_sales();
 
     Ok(())
 }
 
 #[derive(Accounts)]
+#[instruction(amount: String)]
 pub struct StoreSale<'info> {
     /// CHECK: This is the collector keeper, Needs to sign for transfer.
     #[account(
@@ -59,6 +63,13 @@ pub struct StoreSale<'info> {
         bump,
     )]
     store: Account<'info, Store>,
+
+    #[account(
+        mut,
+        seeds = [STORE_SEED, store.key().as_ref(), amount.as_ref()],
+        bump,
+    )]
+    package: Account<'info, StorePackage>,
 
     #[account(mut)]
     payer: Signer<'info>,