+++ anchor/programs/games/src/state/store/mod.rs
@@ -1,0 +1,5 @@
+pub mod store;
+pub mod package;
+
+pub use store::*;
+pub use package::*;